+++ assets/js/admin.js
@@ -1,1 +1,0 @@
-/* Elev8 OS admin JavaScript foundation. */